+++ module-order/build.gradle
@@ -1,0 +1,10 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation(
+            project(':module-menu'),
+            project(':module-ordertable'),
+            project(':module-common')
+    )
+}
