+++ module-order/src/main/java/kitchenpos/application/dto/OrderCreateCommand.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.service.dto;
+package kitchenpos.application.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
