+++ module-order/src/main/java/kitchenpos/application/dto/OrderLineItemCreateCommand.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.service.dto;
+package kitchenpos.application.dto;
 
 public class OrderLineItemCreateCommand {
