+++ module-order/src/main/java/kitchenpos/infra/OrderLineItemRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.infra;
+package kitchenpos.infra;
 
-import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
